# Conflicts:
#	SSPSolution/GraphicsDLL/GraphicsHandler.cpp

+++ SSPSolution/GraphicsDLL/GraphicsHandler.cpp
@@ -598,10 +598,10 @@ int GraphicsHandler::Render(float deltaTime)
 	}*/
 
 	//Go through all components in the root node and render the ones that should be rendered
-	int renderCap = this->m_staticGraphicsComponents.size();
+	size_t renderCap = this->m_staticGraphicsComponents.size();
 	renderCap = this->m_dynamicGraphicsComponents.size();
 	m_shaderControl->SetVariation(ShaderLib::ShaderVariations::Shadow); // render shadows
-	for (size_t i = 0; i < (size_t)renderCap; i++) //FOR EACH NORMAL GEOMETRY
+	for (size_t i = 0; i < renderCap; i++) //FOR EACH NORMAL GEOMETRY
 	{
 		if (this->m_dynamicGraphicsComponents[i]->active)
 		{
@@ -964,7 +964,6 @@ int GraphicsHandler::GenerateOctree()
 		size = size / 2;
 		reachedMaxDepth = !(size > this->m_minSize);
 	}
-
 	//this->m_maxDepth = int((largestSize / this->m_minSize) + 0.5f);
 
 	//Initialize the octree root
